feat(response-compression): switch to synchronous gzip compression
This commit transitions our response compression from an asynchronous method using `TextEncoder` and `TextDecoder` for handling the text, to a synchronous `zlib.gzipSync` function call. It removes the need for the `text-encoding-utf-8` package and simplifies the compression logic within the `Response` class's `compress` method. We're also updating image assets to `.webp` formats in various HTML files, likely as part of a performance optimization strategy for faster loading times and decreased bandwidth usage.

Additionally:
- The `text-encoding-utf-8` package has been removed from the dependencies in both `package.json` and `package-lock.json`.
- Added three new WebP image assets (`banner.webp`, `logo-500.webp`, `logo.webp`).
- Updated image references in `main.html` and `index.html` to point to the new `.webp` images instead of the `.png` versions.